common: Separate atomic operations from fi.h into own header
